move PlayerPair class to own directory
